Use manual updating of versions in l3packages
Using $Id here causes some issues with automated testing (anything
which checks out not using Git not SVN gets "0000/00/00" as a date).
Moreover, I have to hand-edit three of the four files here at every
release in any case, and will add the same interlock to xfrac
(should really be there). As such, there is no additional work for
me in this change.