Largely switch from \ExplFileDate to explicit value
As this is auto-edited by l3build there is no general need to have
a defined variable. In expl3 for the present that is retained: it
leaves \ExplFileDate as a version string for the release as a whole
but is also easier than dropping it for doing the loader/code check.